ccon: Case len to ssize_t in handle_parent
I don't bother checking for overflow here, because ... stupid.

Caught by Splint 3.1.2.

+++ ccon.c
@@ -248,7 +248,7 @@ static int handle_parent(json_t * config, pid_t cpid, int *to_child,
 
 	line = USER_NAMESPACE_MAPPING_COMPLETE;
 	len = strlen(line);
-	if (write(*to_child, line, len) != len) {
+	if (write(*to_child, line, len) != (ssize_t) len) {
 		perror("write to container");
 		return 1;
 	}
